Bug 1347076 - Fix crash in Plugin Kit on rotation with extension displayed
Once upon a time, iOS seemed incapable of properly redisplaying modally presented views on rotation. We would therefore have to remember which view controllers were presented and then re-present them after rotation.
Now this seems to have been fixed, but the old code led to us trying to present modally a view controller that was already presented modally and therefore iOS threw a tantrum and quit.
This seems to be only a problem with UIActivityViewControllers, so we no longer remember the share menu when we display it so we don't try and re-add on rotation.
Tested with iPad Air 2 on iOS 10 & 9.3, iPhone 7 Plus and iPhone 6S Plus 10 & 9.3. The fact that we don't need to have separate code for iOS 10 and iOS 9.x suggests that this was a fix brought in in Swift 3.